+++ tests/testthat/test-corner-cases.R
@@ -5,20 +5,20 @@ test_that("Bug #10 is fixed", {
 
   dates <- c("", "2014-10-29", "2012-12-05")
   pd <- parse_date(dates)
-  pd2 <- c(parse_date(""),
+  pd2 <- cdate(parse_date(""),
            parse_date("2014-10-29"),
            parse_date("2012-12-05")
            )
 
   expect_equal(pd, pd2)
-  
+
 })
 
 test_that("Empty strings are replaced by NAs", {
 
   dates <- c("", "", "")
-  expected <- as.POSIXct(c(NA, NA, NA))
-  actual <- c(parse_date(dates))
+  expected <- .POSIXct(c(NA_real_, NA_real_, NA_real_), "UTC")
+  actual <- parse_date(dates)
 
   expect_equal(expected, actual)
   
@@ -27,8 +27,8 @@ test_that("Empty strings are replaced by NAs", {
 test_that("White-space string is replaced by NAs", {
 
   date <- " "
-  expected <- as.POSIXct(NA)
-  actual <- c(parse_date(date))
+  expected <- .POSIXct(NA_real_, "UTC")
+  actual <- parse_date(date)
 
   expect_equal(expected, actual)
   
@@ -38,7 +38,7 @@ test_that("White-spaces are gracefully removed", {
 
   date <- " 22.2.2222   "
   expected <- parse_date("22.2.2222")
-  actual <- c(parse_date(date))
+  actual <- parse_date(date)
 
   expect_equal(expected, actual)
   
@@ -48,7 +48,7 @@ test_that("White-spaces are gracefully removed", {
 test_that("Non-sensical input is removed", {
   
   date <- "?=)(!$§#$%"
-  expected <- as.POSIXct(NA)
+  expected <- .POSIXct(NA_real_, "UTC")
   actual <- parse_date(date)
   
   expect_equal(expected, actual)
@@ -68,7 +68,7 @@ test_that("Approx parameter is passed on", {
 test_that("zero length input (issue #20)", {
   expect_identical(
     parse_date(character(0)),
-    structure(numeric(0), class = c("POSIXct", "POSIXt"), tzone = ""))
+    structure(numeric(0), class = c("POSIXct", "POSIXt"), tzone = "UTC"))
 
   expect_identical(
     parse_iso_8601(character(0)),